refactor mockWorkspace to TS
