--- airbyte-webapp/src/test-utils/mock-data/mockWorkspace.json
@@ -1,14 +1,0 @@
-{
-  "workspaceId": "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
-  "customerId": "55dd55e2-33ac-44dc-8d65-5aa7c8624f72",
-  "email": "krishna@airbyte.com",
-  "name": "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
-  "slug": "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
-  "initialSetupComplete": true,
-  "defaultGeography": "auto",
-  "displaySetupWizard": false,
-  "anonymousDataCollection": false,
-  "news": false,
-  "securityUpdates": false,
-  "notifications": []
-}
