+++ airbyte-webapp/src/test-utils/mock-data/mockWorkspace.ts
@@ -1,0 +1,15 @@
+import { WorkspaceRead } from "core/request/AirbyteClient";
+
+export const mockWorkspace: WorkspaceRead = {
+  workspaceId: "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
+  customerId: "55dd55e2-33ac-44dc-8d65-5aa7c8624f72",
+  email: "krishna@airbyte.com",
+  name: "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
+  slug: "47c74b9b-9b89-4af1-8331-4865af6c4e4d",
+  initialSetupComplete: true,
+  displaySetupWizard: false,
+  anonymousDataCollection: false,
+  news: false,
+  securityUpdates: false,
+  notifications: [],
+};
